Update the mission
Following the discussion on #122 we've decided to go for a shorter and more concrete version of our mission.

We've decided not to go with a higher level mission because this is actually what we're trying to do and there are so many higher level things we would serve that this might be a very long list.

Furthermore, having a mission that suits what we are doing means we can collect collaborators that want to reach the same goal for different reasons.