Ignore rate limits
Twitter and Github can throw 429 error when too many requests are sent.
This can come from our tests happening concurrently or other testing
happening on the same IP address.

If a rate limiting error is thrown it can be assumed that this
resource exists over that it is broken.

Fixes #695